Portal home dashboard, assignment visibility, cuid-free breadcrumbs
/portal is now the non-member home: a card dashboard of everything a
Dartmouth student can do (hiring application with live cycle/status-aware
CTA, education with enrolled/pending counts and an assignments-due badge)
— designed to grow a card per future surface. The hiring tracker moved to
/portal/hiring; the apply flow's post-submit confetti redirect, the My
Application back-link, and the portal e2e URL assertions follow it, and
the applicant navbar gains Home / Apply / Education.

Assignments stop being buried: creating one notifies every approved
enrollee (in-app deep link for members, email for portal students), the
hub's Assignments tab shows an unsubmitted-count badge, and enrolled
catalog cards (which now link straight to the course hub) carry an
'N assignments due' chip fed by listCatalog.

Breadcrumbs no longer render raw cuids anywhere: any matched route's
handle.breadcrumb can label its own path segment, unlabeled opaque-id
segments (cuids/UUIDs) drop out of the trail (a trailing one falls back
to 'Details'), and the education segments get proper labels.
